\n Application Update Authentication View
In the authentication view, new code has been added to work with csrftoken instead of using a 256-character token generated by the secrets module. The import statement for the secrets module and the code secrets.token_hex(256)—which generated a unique 256-character token—have been removed. Now, the CSRF token is saved in session.session[] using the key 'rT7gM2sP5qW8jN4', simplifying its use in the consumers.py.

Database Configuration

The database connection password for PostgreSQL has been updated with new credentials in the settings.py file.

JavaScript Function Update in Template

In the MultiplayerOnline application, within the buttons_play.html template, the JavaScript function checkMatchStatus has been updated. When searching for a match, a new condition game_status = playing has been added. This allows the script to recognize both playing and completed statuses when searching for matches in the lobbies and retrieving information from the backend. For example:

javascript

if (data.game_status === 'completed' || data.game_status === 'playing') {
    console.log('Match status is: ' + data.game_status);
    clearInterval(interval); // Stop the verification
    window.location.href = '/chess/' + data.id;
} else {
    console.log('Match status is', data.game_status);
    console.log('ID:', data.id);
}

Changes in consumer.py

In the MultiplayerOnline application, the consumer previously accepted the connection and then verified the user. Now, the user is verified before accepting the connection. It is crucial to check for potential bugs if the user is not authenticated.

The code has been updated to work with the new cookie added in the authentication apps, and new conditions have been added for matches that are in progress or playing. Additionally, code has been added to save the chess match status and the FEN (Forsyth-Edwards Notation) format of the board in the database.

Cookie Handling

The code was updated for the 'cookie' because it was realized that the cookie was the same for both players, which was incorrect. I added string 'cookie' to add new code on the future

self.chess_match['white'] = [self.match_query.white_player, 'cookie']
self.chess_match['black'] = [self.match_query.black_player, 'cookie']

Handling Disconnections

When the socket connection is disconnected, without adding self.group_name = 'Disconnecting', a conditional has been added to assign a variable based on whose turn it is:

if chess_match.turn:
    turn = 'white'
else:
    turn = 'black'

Sending var Turn Data

In self.send, new data has been added to send the information to the client to update the status in the templates:

self.send({
    'turn': turn,
    # other data
})

Match Status Update

Code has been added to change the match status from 'Completed' to 'playing' when the match starts.
JavaScript Updates in chess.js

In the MultiplayerOnline application, in Chess.js, a new variable this.chess_board_fen has been added to the Chess class to save the board status on the client. A new method update_chessboard has been added to update the board if the client leaves the page.

To resolve the issue where the opponent could move your piece, the following condition was added at the top of the show_allowed_move method:

javascript

if(this.turn !== this.id){
    return null;
}

WebSocket URL Update in socket.js

The url variable in socket.js has been updated to ensure the WebSocket connection always targets the server, resolving issues in production or when accessing from a device outside localhost:

javascript

const protocol = window.location.protocol === 'https:' ? 'wss:' : 'ws:';
const socketUrl = ${protocol}///ws/Room/;
const socket = new WebSocket(socketUrl);

Views Update in MultiplayerOnline

In the MultiplayerOnline views, the condition and database request in CreateMatch have been updated to include the new status data 'playing'. This ensures that when the script requests to check which lobbies are completed, waiting, or playing, it retrieves accurate information from the database.